% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateSitesInRepeat.R
\name{annotateSitesInRepeat}
\alias{annotateSitesInRepeat}
\title{Obtain the position, type, name of the repeat elements that contain a list of
sites}
\usage{
annotateSitesInRepeat(
  mySiteList,
  AnnotationData,
  SiteLocForDistance = c("middle", "start", "end"),
  RepeatLocForDistance = c("middle", "start", "end"),
  output = c("hits", "all"),
  ignore.strand = TRUE,
  ...
)
}
\arguments{
\item{mySiteList}{A \link[GenomicRanges:GRanges-class]{GRanges} object}

\item{AnnotationData}{A \link[GenomicRanges:GRanges-class]{GRanges} or}

\item{SiteLocForDistance}{middle, start, or end}

\item{RepeatLocForDistance}{middle, start, or end}

\item{output}{either "all" or "hits" only}

\item{ignore.strand}{whether or not strand gets ignored}
}
\description{
Obtain the position, type, name of the repeat elements that contain a list of
sites
}
\examples{


    ## example 1: annotate mySiteList by Repeat.hg38, and output all sites
    library(TECookbook)
    data(Reoeats.Human.hg38)
    annotateSite = annotateSitesInRepeat(mySiteList, AnnotationData=Repeats.Human.hg38,
                   SiteLocForDistance = "middle", RepeatLocForDistance = "middle",
                   output = "all")

    ## example 2: annotate mySiteList (GRanges) and only output hits
    annotateSite = annotateSitesInRepeat(mySiteList, AnnotationData=Repeats.Human.hg38,
                   SiteLocForDistance = "middle", RepeatLocForDistance = "middle",
                   output = "hits")

}
