% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildAnnotation.R
\name{buildAnnotation}
\alias{buildAnnotation}
\title{Function to convert RepeatMasker data to Genomic Range for annotating ChIP-seq peaks or binding sites}
\usage{
buildAnnotation(file)
}
\arguments{
\item{file}{RepeatMasker file used for extracting repeat name and genomic region, e.g., hg38.fa.out}
}
\description{
Function to convert RepeatMasker data to Genomic Range for annotating ChIP-seq peaks or binding sites
}
\examples{
         ##Example:
         Repeats.Human.hg38 = buildAnnotation("hg38.fa.out")
}
