% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildChain.R
\name{buildChain}
\alias{buildChain}
\title{Construct Chain file and Repeat sizes file for liftOver operation between genome coordinates and repeat coordinates}
\usage{
buildChain(
  alignment = "hg38.fa.align",
  chainFile = "Hg38ToRepeat.over.chain",
  sizeFile = "Repeat.sizes",
  include = NULL
)
}
\arguments{
\item{alignment}{RepeatMasker alignment file used for extracting genomic region, e.g., hg38.fa.align}

\item{chainFile}{Name of liftOver chain file to be written}

\item{sizeFile}{Name of repeat sizes file to be written}

\item{include}{whether or not only build chain file for select repeat family; if NULL, output all repeat instances;}
}
\description{
Construct Chain file and Repeat sizes file for liftOver operation between genome coordinates and repeat coordinates
}
\examples{
         ##Example 1: Construct chain file for all repeats in hg38 genome
         buildChain(alignment = "hg38.fa.align", chainFile = "Hg38ToRepeat.over.chain", sizeFile = "hg38.Repeat.sizes")
         ##Example 2: Construct chain file for THE1A repeats only in hg38 genome
         buildChain(alignment = "hg38.fa.align", chainFile = "Hg38ToTHE1A.over.chain", sizeFile = "THE1A.Repeat.sizes", include="THE1A")
}
