% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liftOut.R
\name{liftOut}
\alias{liftOut}
\title{Function to extract all genomic regions that maps to some specific locus within a reference repeat element}
\usage{
liftOut(alignment, Repeat, start_pos, end_pos, RepeatID = NULL)
}
\arguments{
\item{alignment}{RepeatMasker alignment file used for extracting genomic region, e.g., hg38.fa.align}

\item{Repeat}{Name of Repeat to be extracted}

\item{start_pos}{Starting position of extracted locus within repeat consensus sequence, defined in Dfam database}

\item{end_pos}{End position of extracted locus within repeat consensus sequence, defined in Dfam database}

\item{RepeatID}{Extract regions only within defined RepeatIDs; If NULL, no such restriction.}
}
\description{
Function to extract all genomic regions that maps to some specific locus within a reference repeat element
}
\examples{
         ##Example 1: lift out all sites from THE1B repeats in hg38 genome
         sites = liftOut(alignment = "hg38.fa.align", Repeat = "THE1B", start_pos = 202, end_pos = 221)
         ##Example 2: lift out sites from L1P1_orf2 repeats with repeatID 1000 and 1005
         sites = liftOUt(alignment = "hg38.fa.align", Repeat = "L1P1_orf2", start_pos = 1000, end_pos = 1020, RepeatID = c(1000, 1005))
}
